fix cast/conversion warnings & API CHANGE for HL encoder
  * in high level encoder API: disp is now SIGNED.

  * Found using gcc5, gcc8 and clang(xcode10) -Wconversion
    -Wsign-conversion (disabled for library build since about 150
    warnings remain, some in generated code)

  * For examples, pass --example-flags '-Wconversion -Wsign-conversion' to the build.

(cherry picked from commit 0434efd5cb297c4be873d3fd3f14c85b79a67ce2)

+++ examples/xed-symbol-table.c
@@ -137,9 +137,9 @@ int xed_disassembly_callback_function(
         char* symbol  = get_symbol(symbol_address, caller_data);
         if (symbol) {
             if (xed_strlen(symbol) < buffer_length)
-                xed_strncpy(symbol_buffer, symbol, buffer_length);
+                xed_strncpy(symbol_buffer, symbol, (int)buffer_length);
             else {
-                xed_strncpy(symbol_buffer, symbol, buffer_length-1);
+                xed_strncpy(symbol_buffer, symbol, (int)(buffer_length-1));
                 symbol_buffer[buffer_length-1]=0;
             }
             *offset = address - symbol_address;
